requirement.txt
asgiref==3.7.2
boto3==1.34.60
botocore==1.34.60
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==41.0.1
Django==4.2.2
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-environ==0.10.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
et-xmlfile==1.1.0
idna==3.6
jmespath==1.0.1
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.0.7
